Renames to fix repository case issue
