+++ src/components/home-component.vue
